docs: revamp readme and restructure project organization
