+++ src/matrix_operations/minor_matrices.py
